// src/enum/screen-step.enum.ts
export enum ScreenStepEnum {
  HOME_SCREEN_INITIAL = 'HOME_SCREEN_INITIAL',

  NOT_FOUND_SCREEN_INITIAL = 'NOT_FOUND_SCREEN_INITIAL',

  TASK_SCREEN_INITIAL = 'TASK_SCREEN_INITIAL',
  TASK_SCREEN_DOWNLOADING = 'TASK_SCREEN_DOWNLOADING',
  TASK_SCREEN_ENTER_NEW_TASK = 'TASK_SCREEN_ENTER_NEW_TASK',
  TASK_SCREEN_PROCESSING_TASK_ADDITION = 'TASK_SCREEN_PROCESSING_TASK_ADDITION',
  TASK_SCREEN_SHOW_DETAIL = 'TASK_SCREEN_SHOW_DETAIL',
  TASK_SCREEN_ENTER_SUPERVISOR_PASSWORD = 'TASK_SCREEN_ENTER_SUPERVISOR_PASSWORD',
  TASK_SCREEN_PROCESSING_STATUS_CHANGE = 'TASK_SCREEN_PROCESSING_STATUS_CHANGE',
  TASK_SCREEN_PROCESSING_REQUEST_FOR_NEW_TASKS = 'TASK_SCREEN_PROCESSING_REQUEST_FOR_NEW_TASKS',
  TASK_SCREEN_OPEN_FLOAT_MENU = 'TASK_SCREEN_OPEN_FLOAT_MENU',
}